requirements.txt
Flask>=0.12.3
Jinja2>=2.9.5
MarkupSafe==0.23
Werkzeug>=0.14
click==6.7
itsdangerous==0.24
requests==2.20.0
wsgiref==0.1.2 ; python_version < '3.2'
plaid-python>=5.0.0, < 6.0.0
